Saturn Satisfier initial emulation
A somewhat UNIX-centric implementation for now. Defaults to using
/tmp/satisfier as the served-up filesystem root. No config options for
this yet.

The architecture is a bit weird because it reuses the main command
dispatching code from the Satisfier in order to maximise compatibility.
This dispatch code lives in cdb_command_impl.c, which is separately
licensed.

The filesystem interface code lives in fffake.c, and essentially
emulates the API of FatFS. This will need some work to bring it up to
scratch on Windows.

This is a moving target - the device firmware is still changing too. As
yet there's no way to enter/leave command mode (which will conceal the
added CDB commands from non-aware application software). A new CD
backend to stream sectors from the Satisfier root will follow.